app/requirements.txt
flask==3.0.3
gunicorn==22.0.0
werkzeug==3.0.3
itsdangerous==2.2.0
markupsafe==2.1.5
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
cryptography==42.0.8
flask-login==0.6.3
flask-wtf==1.2.1
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
email_validator==2.2.0
